+++ CHANGELOG.md
@@ -14,7 +14,8 @@
 * Added support to apply gradient colors to theme [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
 * Updated [Browserslist](./src/browserslist) configurations that automatically add autoprefixer to ensure compatibility with different browser and browser versions. Excluded browser <= IE 10 [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
 * Implemented loading indicator to show loading progress between route changes using [Angular Material Progress Bar](https://material.angular.io/components/progress-bar/overview) and Angular Router Events [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
-* Implemented parallax effect using CSS [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
+* Implemented parallax effect using SCSS [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
+* Implemented shrinkable second row header menu bar on scrolling down/up using [Angular Material CDK](https://material.angular.io/cdk/scrolling/api) `ScrollingModule` and SCSS [[PR 4](https://github.com/kumaran-is/ngrx-web-starter1/pull/4)]
 
 #### Breaking Changes
 * NA